nlkklnnlknkl

+++ src/main/resources/db/migration/V1__Create_Role_Table.sql
@@ -1,6 +1,5 @@
-
-
 CREATE TABLE role (
-                      id BIGINT PRIMARY KEY ,
+                      id SERIAL PRIMARY KEY,
                       name VARCHAR(50) NOT NULL
 );
+
